pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-agent-starter"
version = "0.1.0"
description = "A local starter project for LLM and agentic AI concepts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
ai-starter = "llm_agent_starter.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
